Moved new js files in a new folder called helpers
