+++ frontend/src/helpers/DetectBrowser.js
